+++ scripts/creation.js
@@ -49,7 +49,7 @@ const create = async () => {
       state order_states NOT NULL,
       order_date TIMESTAMP NOT NULL DEFAULT (now() AT TIME ZONE 'UTC'),
       delivery_date TIMESTAMP NOT NULL DEFAULT (now() AT TIME ZONE 'UTC'),
-      cart BOOLEAN NOT NULL DEFAULT false
+      order_total DECIMAL
     );
 
     CREATE TABLE IF NOT EXISTS cards (
@@ -65,8 +65,6 @@ const create = async () => {
     CREATE TABLE IF NOT EXISTS products_orders (
       product_id uuid references products(id),
       order_id uuid references orders(id),
-      order_id uuid references orders(id),
-      order_total DECIMAL not null
     );
     `);
     console.info("> creation done! 🚀");
