+++ services/index.js
@@ -3,7 +3,7 @@ const router = require("express").Router();
 module.exports = (db) => {
   router.use("/auth", require("./auth")(db));
   router.use("/products", require("./products")(db));
-  router.use("/cart", require("./cart")(db));
+  router.use("/cart", require("./products")(db));
 
   return router;
 };
